refactor(server): move code blocks to functions
refactor: change server_age variable type

+++ headers/server.h
@@ -74,7 +74,7 @@ void remove_first_client();
 /* SERVER */
 void terminate_connection(const int connfd);
 off_t *get_authorization_end_at();
-void get_server_time(time_t *server_start_at, uint64_t *server_age);
+void get_server_time(time_t *server_start_at, uint32_t *server_age);
 void start_server(struct Configuration *config);
 struct Configuration *get_server_configuration();
